Refactored ListGroup and updated some unit tests
