+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "lib-react-ui",
-  "version": "1.0.3",
+  "version": "1.0.4",
   "description": "Reusable React UI Components",
   "main": "dist/bundle.cjs.js",
   "module": "dist/bundle.esm.js",
@@ -28,7 +28,6 @@
     "prop-types": "^15.8.1"
   },
   "devDependencies": {
-    "@a110/storybook-expand-all": "^1.1.1",
     "@babel/preset-env": "^7.22.5",
     "@babel/preset-react": "^7.22.5",
     "@babel/preset-typescript": "^7.22.5",
