+++ src/components/ListGroup/Group/ListGroup.types.ts
@@ -1,0 +1,7 @@
+import { ChildNodes } from "../../../globals";
+
+export interface ListGroupProps {
+  id?: string;
+  children?: ChildNodes;
+  className?: string;
+}
