+++ src/components/ListGroup/Item/ListGroupItem.types.ts
@@ -1,0 +1,7 @@
+import { ChildNodes } from "../../../globals";
+
+export interface ListGroupItemProps {
+  id?: string;
+  children?: ChildNodes;
+  className?: string;
+}
