--- src/components/ListGroup/ListGroup.types.ts
@@ -1,13 +1,0 @@
-import { ChildNodes } from "../../globals";
-
-export interface ListGroupProps {
-  id?: string;
-  children?: ChildNodes;
-  className?: string;
-}
-
-export interface ListGroupItemProps {
-  id?: string;
-  children?: ChildNodes;
-  className?: string;
-}
