+++ src/components/ListGroup/index.ts
@@ -1,1 +1,2 @@
-export * from "./ListGroup";
+export * from "./Group";
+export * from "./Item";
